wallet improve

+++ app/Http/Controllers/ApiController.php
@@ -76,11 +76,11 @@ class ApiController extends Controller
     {
         $model = $this->repositry->getByID($id);
 
-        if ($model) {
-            return $this->returnData('data', new $this->resource( $model ), __('Get  succesfully'));
+        if (!$model) {
+            return $this->returnError(__('Sorry! Failed to get !'));
         }
+        return $this->returnData('data', new $this->resource( $model ), __('Get  succesfully'));
 
-        return $this->returnError(__('Sorry! Failed to get !'));
     }
 
     /**
@@ -91,6 +91,12 @@ class ApiController extends Controller
      */
     public function delete($id)
     {
+        $model = $this->repositry->getByID($id);
+
+        if (!$model) {
+            return $this->returnError(__('Sorry! Failed to get !'));
+        }
+
         $this->repositry->deleteByID($id);
 
         return $this->returnSuccessMessage(__('Delete succesfully!'));
